install c++ deps

+++ provision.sh
@@ -1,12 +1,12 @@
 yum -y install gcc automake autoconf libtool make
-yum install gcc gcc-c++
+yum -y install gcc
+yum -y install gcc-c++
 
 
 cd /usr/local/src
 
-#curl -k -s https://jaist.dl.sourceforge.net/project/pcre/pcre/8.37/pcre-8.37.tar.gz -o pcre.tar.gz
-#curl -k -s https://nchc.dl.sourceforge.net/project/pcre/pcre/8.34/pcre-8.34.tar.gz -o pcre.tar.gz
-curl -k https://jaist.dl.sourceforge.net/project/pcre/pcre/8.34/pcre-8.34.tar.gz -o pcre.tar.gz
+#for url rewrite
+if [ ! -e pcre.tar.gz ] ; then curl -k https://jaist.dl.sourceforge.net/project/pcre/pcre/8.34/pcre-8.34.tar.gz -o pcre.tar.gz ; fi
 tar -zxvf pcre.tar.gz
 cd pcre-8.34
 ./configure
@@ -14,7 +14,7 @@ make
 make install
 
 cd /usr/local/src
-curl -k https://nchc.dl.sourceforge.net/project/libpng/zlib/1.2.11/zlib-1.2.11.tar.gz -o zlib.tar.gz
+if [ ! -e zlib.tar.gz ] ; then curl -k https://nchc.dl.sourceforge.net/project/libpng/zlib/1.2.11/zlib-1.2.11.tar.gz -o zlib.tar.gz ; fi
 tar -zxvf zlib.tar.gz
 cd zlib-1.2.11/
 ./configure
@@ -22,7 +22,7 @@ make
 make install
 
 cd /usr/local/src
-curl -k https://www.openssl.org/source/openssl-1.0.1t.tar.gz -o openssl.tar.gz
+if [ ! -e openssl.tar.gz ] ; then curl -k https://www.openssl.org/source/openssl-1.0.1t.tar.gz -o openssl.tar.gz ; fi
 tar -zxvf openssl.tar.gz
 cd openssl-1.0.1t/
 ./config shared zlib  --prefix=/usr/local/openssl   
@@ -30,7 +30,7 @@ make
 make install
 
 cd /usr/local/src
-curl http://nginx.org/download/nginx-1.4.2.tar.gz -o nginx.tar.gz
+if [ ! -e nginx.tar.gz ] ; then curl http://nginx.org/download/nginx-1.4.2.tar.gz -o nginx.tar.gz ; fi
 tar -zxvf nginx.tar.gz
 cd nginx-1.4.2/
 ./configure --sbin-path=/usr/local/nginx/nginx \
